refactor: split CartPreview in separate components
