+++ src/types/index.ts
@@ -5,3 +5,4 @@ export * from "./menu.types";
 export * from "./purchase.types";
 export * from "./user.types";
 export * from "./cart.types";
+export * from "./cart.helpers";
